% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert.urine.to.exposure}
\alias{convert.urine.to.exposure}
\title{convert.urine.to.exposure}
\usage{
convert.urine.to.exposure(
  urine.conc,
  urine.vol.per.day = 1.4,
  average.creatinine = 122.6
)
}
\arguments{
\item{urine.conc}{Urine concentration in ug/g creatinine/kg bodyweight}

\item{urine.vol.per.day}{Urine volume per day in liters.  Default is 1.4 L.}

\item{average.creatinine}{Average creatinine concentration in mg/DL.
Default is 122.6 mg/DL.}
}
\value{
Exposure dose in mg/kg bodyweight/day
}
\description{
Converts urine concentration in ug/g creatinine/kg bodyweight
to mg/kg bodyweight/day
}
