% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{creatinine}
\alias{creatinine}
\title{creatinine}
\usage{
creatinine(newdata)
}
\arguments{
\item{newdata}{A data.frame with individual-specific data including the
following data columns: RIAGENDR, RIDRETH1, BMXWT, RIDAGEYR}
}
\value{
Estimated daily creatinine excretion in mg creatinine/day
}
\description{
Estimates daily creatinine excretion in mg creatinine/day
}
