% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examine_distribution.R
\name{examine_error}
\alias{examine_error}
\title{examine_error}
\usage{
examine_error(
  Measured,
  codes_file,
  data_path = ".",
  save_directory = ".",
  cores = NULL
)
}
\arguments{
\item{Measured}{A data frame containing a row for each metabolite with columns
providing the chemical identifiers, log geometric means, subpopulation,
and other data carried on from the codes file.}

\item{codes_file}{Manually created xls or xlsx file containing 3 sheets:  1. NHANES
chemicals to include (with identifier, code, file, demographic, and units),
2. Associated weights, filenames, and column names associated with each phase used,
and 3. Parent-metabolite map containing chemical identifiers and molecular weights.}

\item{data_path}{String providing the path to the raw data.  Default is ".",
the direct output from running get_NHANES_data would be "rawData", resulting
in e.g. ./rawData/1999-2000.}

\item{save_directory}{String giving the path of where to save the output table
and plot.}

\item{cores}{Allows for the call to doplots(), which is called by examine_error, to be run in parallel
for each unique chemical in Measured. Default is NULL, which uses a traditional for loop.}
}
\value{
Measured: same as the input, but with 4 additional columns.
1. Prabove: fraction of measurements above the LOD.
2. Rightseg: right interval of curve around expected
mean of distribution
3. Leftseg: left interval of curve around expected
mean of distribution
4. PosteriorShape: either uniform or normal. Decided
based on Prabove and the standard error distribution.
}
\description{
Examines the standard errors of the maximum likelihood estimates and
the number of measurements below the LOD.  If either of those
}
