% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromMolar.R
\name{fromMolar}
\alias{fromMolar}
\title{fromMolar}
\usage{
fromMolar(
  SUBPOP,
  Measured,
  pred.data,
  nhanesdata,
  out.coda3R,
  doplot = TRUE,
  codes_file,
  save_directory = "."
)
}
\arguments{
\item{SUBPOP}{The subpopulation name for which we want to look at convergence}

\item{Measured}{Output from the examine_error function (saved as
NewMeasured), with the last column being PosteriorShape.
Same as input for fitOnlyP}

\item{pred.data}{Output from the readNHANES function, saved in MCMCdata file}

\item{nhanesdata}{Output from fitOnlyP. A list containing all the data needed to run the jags model.
Used as input to the fromMolar function to assign chemical
information to the final estimates.}

\item{out.coda3R}{Output from fitOnlyP. A MCMC list object.}

\item{doplot}{Logical input to indicate whether a plot should be generated and saved.
Default is TRUE.}

\item{codes_file}{Manually created xls or xlsx file containing 3 sheets:  1. NHANES
chemicals to include (with identifier, code, file, demographic, and units),
2. Associated weights, filenames, and column names associated with each phase used,
and 3. Parent-metabolite map containing chemical identifiers and molecular weights.}

\item{save_directory}{String specifying the directory in which to save the plot and results.}
}
\value{
lPsampsgm: data frame whose rows are the parent chemicals, columns are
the different subpopulations, and the values are the exposure
estimates in mg/kg bodyweight/day for that cheimcal in that
subpopulation.
Measured: same as the input to fitOnlyP, reduced to rows matching the
input subpopulation. Geometric means of the metabolites that
are created from the parents in lPsampsgm.
Plot: ggplot result of comparison of metabolite estimates. View with print(Plot).
}
\description{
This is generic code for checking the results of the MCMC analyses of
the NHANES-based modeling and for converting the estimates from units
of nmole / kg bodyweight / day to units appropriate for exposure
mg / kg bodyweight / day.
We check convergence first, using heidel.diag() and gelman.diag
on the three chain results.  Check both lP and lU (for OnlyP).
Then, compare the lU estimates to the values of Measured$loggm.  Just a plot
should do. This is not a particularly rigorous test, but if it fails,
certainly we have problems.
Finally, do the conversion, and save.
}
