% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{gcens}
\alias{gcens}
\title{gcens}
\usage{
gcens(x, mean, lsd)
}
\arguments{
\item{x}{Log (ln) of measurement value}

\item{mean}{Unnamed x}

\item{lsd}{Log (ln) of standard deviation}
}
\value{
???
}
\description{
Derivative of the loglike function.  Obtain the gradient
to help in calculating variance estimates.
}
