% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{geometric.mean.to.IUR}
\alias{geometric.mean.to.IUR}
\title{geometric.mean.to.IUR}
\usage{
geometric.mean.to.IUR(x)
}
\arguments{
\item{x}{Geometric mean production volumne}
}
\value{
IUR production volume
}
\description{
Obtains IUR production volume code from geometric mean production
volume
}
