% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getDesign}
\alias{getDesign}
\title{getDesign}
\usage{
getDesign(
  demof = "DEMO_F.XPT",
  chemdtaf,
  measurehead = "URX",
  measuretail = NULL,
  lodindhd = "URD",
  lodtail = "LC",
  seq = "SEQN",
  demoageyr = "RIDAGEYR",
  demogendr = "RIAGENDR",
  demoeth = "RIDRETH1",
  PSU = "SDMVPSU",
  STRA = "SDMVSTRA",
  chem2yrwt = NULL,
  nm = NULL,
  creatinine = "URXUCR",
  CreatFun = NULL,
  creatfile = NULL,
  urinerate = NULL,
  urinefile = NULL,
  bodywtfile = NULL,
  bodywt = "BMXWT",
  bodywtcomment = "BMIWT",
  bodymassindex = "BMXBMI",
  MECwt = "WTMEC2YR",
  codes_table = NULL,
  code = NULL
)
}
\arguments{
\item{demof}{String containing name of SAS transport file containing
demographic data}

\item{chemdtaf}{String containing name of SAS transport file containing
the urine concentrations for a group of chemicals.
presumed to have cols SEQN, WTSB2YR, and pairs of column for each chemical: one
for the measurement and one for an indicator for < LOD.  The value in the measurement
when the measurement is < LOD is LOD/sqrt(2), rounded to about 2 digits.}

\item{measurehead}{String matching the initial part of the measurement variable
names.  The defaults are correct for files for laboratory measurement of
products in urine.}

\item{measuretail}{Strings matching the final part of the measurement variable
names.  The defaults are correct for files for laboratory measurement of
products in urine.}

\item{lodindhd}{Regular expression for matching the 'comment' fields (lod indicator), initial part}

\item{lodtail}{Regular expressions for matching the 'comment' fields (lod indicator), final part}

\item{seq}{ID, used to merge different files (assumed to be the same variable name across files)}

\item{demoageyr}{Name of the variable giving age in years in the demographic data file.}

\item{demogendr}{Name of the variable giving gender in the demographic data file.}

\item{demoeth}{Name of the variable giving ethnicity in the demographic data file.}

\item{PSU}{Name of the variable in the demographic data file giving the sampling unit}

\item{STRA}{Name of the variable in the demographic data file giving the stratum for each observation}

\item{chem2yrwt}{Variable in chemdataf giving the sampling weight to be used for those data}

\item{nm}{Column name, also chemical code, for the chemical to get design for}

\item{creatinine}{Name of the variable in the file named by chemdataf that contains the creatinine concentration
in each urine sample}

\item{CreatFun}{Function to compute median or random daily estimates of creatinine excretion.  Used to convert
outputs scaled by creatinine concentration to daily excretion rates. The argument is a data frame with variables:
RIAGENDR, RIDRETH1, BMXWT, and RIDAGEYR}

\item{creatfile}{Name of file with creatinine measurements}

\item{urinerate}{Variable in urinefile giving the rate of urine production}

\item{urinefile}{File containing basic information about urine rate and volume.  ONly relevant to 2009 and
later}

\item{bodywtfile}{File giving the body weight information}

\item{bodywt}{Variable giving bodyweight in kg}

\item{bodywtcomment}{Variable flagging special cases for bodyweight}

\item{bodymassindex}{Code for BMI information in the NHANES file.  Default is "BMXBMI".}

\item{MECwt}{Variable giving the weight to be used when analyzing a full MEC variable}

\item{codes_table}{R dataframe of the first sheet of the codes_file in the readNHANES() function}

\item{code}{Information for adding chemical names and CAS numbers to the output.  Is a list with elements
table: a character matrix or data frame with character (not
factor) elements giving the names NHANES uses for
chemical measures, the chemical names, and the CAS
numbers as strings.
codename: the name in table of the variable that contains the codes NHANES uses
chemame: tha name in table of the variable containing the chemical names
CAS: either NULL or the name in table of the variable containing the CAS numbers}
}
\value{
A design object
}
\description{
Obtain the design object for a given name nm and subpop.  Same as the first part of
getNhanesQuantiles.
}
