% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNhanesQuantiles}
\alias{getNhanesQuantiles}
\title{getNhanesQuantiles}
\usage{
getNhanesQuantiles(
  demof = "DEMO_F.XPT",
  chemdtaf,
  measurehead = "URX",
  measuretail = NULL,
  lodindhd = "URD",
  lodtail = "LC",
  seq = "SEQN",
  demoageyr = "RIDAGEYR",
  demogendr = "RIAGENDR",
  demoeth = "RIDRETH1",
  PSU = "SDMVPSU",
  STRA = "SDMVSTRA",
  chem2yrwt = NULL,
  chemvars,
  creatinine = "URXUCR",
  CreatFun = NULL,
  creatfile = NULL,
  urinerate = NULL,
  urinefile = NULL,
  bodywtfile = NULL,
  bodywt = "BMXWT",
  bodywtcomment = "BMIWT",
  bodymassindex = "BMXBMI",
  MECwt = "WTMEC2YR",
  Q = c(50, 75, 90, 95),
  lognormfit = TRUE,
  plot = NULL,
  code = NULL,
  LODfilter = TRUE,
  MaximumAge = 150
)
}
\arguments{
\item{demof}{String containing name of SAS transport file containing
demographic data}

\item{chemdtaf}{String containing name of SAS transport file containing
the urine concentrations for a group of chemicals.
presumed to have cols SEQN, WTSB2YR, and pairs of column for each chemical: one
for the measurement and one for an indicator for < LOD.  The value in the measurement
when the measurement is < LOD is LOD/sqrt(2), rounded to about 2 digits.}

\item{measurehead}{String matching the initial part of the measurement variable
names.  The defaults are correct for files for laboratory measurement of
products in urine.}

\item{measuretail}{Strings matching the final part of the measurement variable
names.  The defaults are correct for files for laboratory measurement of
products in urine.}

\item{lodindhd}{Regular expression for matching the 'comment' fields (lod indicator), initial part}

\item{lodtail}{Regular expressions for matching the 'comment' fields (lod indicator), final part}

\item{seq}{ID, used to merge different files (assumed to be the same variable name across files)}

\item{demoageyr}{Name of the variable giving age in years in the demographic data file.}

\item{demogendr}{Name of the variable giving gender in the demographic data file.}

\item{demoeth}{Name of the variable giving ethnicity in the demographic data file.}

\item{PSU}{Name of the variable in the demographic data file giving the sampling unit}

\item{STRA}{Name of the variable in the demographic data file giving the stratum for each observation}

\item{chem2yrwt}{Variable in chemdataf giving the sampling weight to be used for those data}

\item{chemvars}{Variables in chemdataf that are to be analyzed}

\item{creatinine}{Name of the variable in the file named by chemdataf that contains the creatinine concentration
in each urine sample}

\item{CreatFun}{Function to compute median or random daily estimates of creatinine excretion.  Used to convert
outputs scaled by creatinine concentration to daily excretion rates. The argument is a data frame with variables:
RIAGENDR, RIDRETH1, BMXWT, and RIDAGEYR}

\item{creatfile}{Name of file with creatinine measurements}

\item{urinerate}{Variable in urinefile giving the rate of urine production}

\item{urinefile}{File containing basic information about urine rate and volume.  Only relevant to 2009 and
later}

\item{bodywtfile}{File giving the body weight information}

\item{bodywt}{Variable giving bodyweight in kg}

\item{bodywtcomment}{Variable flagging special cases for bodyweight}

\item{bodymassindex}{Code for BMI information in the NHANES file.  Default is "BMXBMI".}

\item{MECwt}{Variable giving the weight to be used when analyzing a full MEC variable}

\item{Q}{Vector of percentiles to estimate (e.g. 50 = 50th percentile)}

\item{lognormfit}{Logical, fit a lognormal distribution to data (including the subpopulations)}

\item{plot}{If not NULL, the name to give a pdf file for plotting CDFs, empirical and lognormal
from parameter estimates}

\item{code}{Information for adding chemical names and CAS numbers to the output.  Is a list with elements
table: a character matrix or data frame with character (not
factor) elements giving the names NHANES uses for
chemical measures, the chemical names, and the CAS
numbers as strings.
codename: the name in table of the variable that contains the codes NHANES uses
chemame: tha name in table of the variable containing the chemical names
CAS: either NULL or the name in table of the variable containing the CAS numbers}

\item{LODfilter}{should the estimated quantiles be filtered so they are NA for values < LOD?  This
is for matching NHANES summary reports.  Only relevant to quantile estimates and their
confidence limits}

\item{MaximumAge}{What is the oldest age in years to consider?}
}
\value{
A dataframe with columns:
NHANEScode: nhanes variable name
Chem: chemical name (if !is.null(code) in arguments)
CAS: CAS number (if !is.null(code) and !is.null(code$CAS))
subpop: subpopulation identifier: Total, Male, Female, 0 - 5, 6 - 11
years, 12 - 19 years, 20 - 65 years, 66 years and older,
ReproAgeFemale, BMI <= 30, BMI > 30
then triplets of variables for each quantile requested.  For example, if Q is c(50, 90),
then Q_50, Q_50_lcl, Q_50_ucl, Q_90, Q_90_lcl, Q_90_ucl
LOD: the limit of detection for each variable
if lognormfit, then
loggm: natural log of the geometric mean (R plnorm parameter meanlog)
loggm_se: standard error of the estimate of loggm; NA if not
computable (in which case, probably loggm and lsdlog are not unique).
lsdlog: the natural log of the standard deviation of the log-transformed variable
(the natural logarithm of the R plnorm parameter sdlog).
lsdlog_se: the standard error of the estimate of lsdlog.  See loggm_se for meaning of NA.

In addition, if a value is given for plot, a pdf file is produced
with estimated population cdfs (i.e., taking account of the NHANES
sampling design), with superimposed quantile estimates and
confidence limits.  If lognormfit is true, then the fitted curve is
superimposed.
}
\description{
Population Quantiles, MLEs of lognormal population parameters
w/standard errors, and comparisons of fitted lognormal curve to
empirical CDF for data.
}
\details{
This code attempts to reproduce the Phthalate quantiles and
confidence limits reported in 'Fourth National Exposure Report,
Updated Tables, March 2013', estimate parameters for an assumed
lognormal population distribution of the urinary metabolites in
mg/day excretion rates, and compare the estimated lognormal
distribution to the estimated population CDF for the sample data,
all using proper adjustments for the NHANES sampling design.

Function for reading relevant files, merging them, and estimating
quantiles for urine concentrations
}
\examples{
# indata <- c(
# "Mono(carboxyoctyl) phthalate;URXCOP
# Mono (carboxynonyl) phthalate;URXCNP
# Mono-n-butyl phthalate;URXMBP
# Mono-isobutyl phthalate;URXMIB
# Mono-n-methyl phthalate;URXMNM
# Mono-benzyl phthalate;URXMZP")
# con <- textConnection(indata)
# tbl <- read.delim(con, sep = ";", header = FALSE, col.names = c("Chem", "varname"),
#                  stringsAsFactors = FALSE)
# close(con)

# inpath <- file.path(find.package("bayesmarker"), "extdata")
# phthalatesQ <- getNhanesQuantiles(demof = file.path(inpath, "DEMO_F.XPT"),
# chemdtaf = file.path(inpath, "PHTHTE_F.XPPPPT"), lognormfit = TRUE, plot = "phthalates.pdf",
# code = list(table = tbl, chemname = "Chem", codename = "varname", CAS = NULL))

}
