% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modparms}
\alias{modparms}
\title{Model inputs for the creatinine function}
\format{
An object of class \code{numeric} of length 13.
}
\usage{
modparms
}
\description{
Mean daily creatinine values for each subpopulation
}
\keyword{datasets}
