% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimates.R
\name{plot_subpops}
\alias{plot_subpops}
\title{plot_subpops}
\usage{
plot_subpops(
  path_to_lPs = NULL,
  SUBPOP = "all",
  save_output = FALSE,
  save_directory = ".",
  print_plot = FALSE
)
}
\arguments{
\item{path_to_lPs}{String representing the location of the lPsamps outputs.
All subpopulations that are desired in the plot need to
be within the directory provided.}

\item{SUBPOP}{A character vector of the subpopulations to be plotted.  These
must exactly match the subpopulations contained in the original
codes_file.}

\item{save_output}{Set to TRUE to save the results. Note that the default
save location is the current working directory.}

\item{save_directory}{String giving the path of where to save the output table
and plot.}

\item{print_plot}{Boolean input. Default is FALSE. Set to TRUE to plot in the
current R session.}
}
\value{
exposureDF:  data.frame with rows representing the chemicals and columns
representing the different subpopulations
Plot: The plot is printed for viewing, but you can choose where to save
it or view it again using print(p)
}
\description{
Plot the median exposure estimate for all subpopulations of each parent
chemical that have calculated exposure estimates.  Sorted from highest
to lowest predicted exposure.
}
