% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UseTable}
\alias{UseTable}
\title{Chemical use information for about 530,000 chemicals}
\format{
A data frame with 532,414 rows and 20 variables
\describe{
\item{CASRN}{chemical CAS number}
\item{COMMON.NAME}{chemical name}
\item{TXCODE}{??}
\item{ANTIMICROBIAL}{is the chemical used this way, 1 for yes, 0 for no}
\item{CHEMICAL.INDUSTRIAL.PROCESS}{is the chemical used this way, 1 for yes, 0 for no}
\item{CHEMICAL.WARFARE}{is the chemical used this way, 1 for yes, 0 for no}
\item{COLORANT}{is the chemical used this way, 1 for yes, 0 for no}
\item{CONSUMER.USE}{is the chemical used this way, 1 for yes, 0 for no}
\item{FERTILIZER}{is the chemical used this way, 1 for yes, 0 for no}
\item{FOOD.ADDITIVE}{is the chemical used this way, 1 for yes, 0 for no}
\item{FRAGRANCE}{is the chemical used this way, 1 for yes, 0 for no}
\item{HERBICIDE}{is the chemical used this way, 1 for yes, 0 for no}
\item{PERSONAL.CARE.PRODUCT}{is the chemical used this way, 1 for yes, 0 for no}
\item{PESTICIDE}{is the chemical used this way, 1 for yes, 0 for no}
\item{PETROCHEMICAL}{is the chemical used this way, 1 for yes, 0 for no}
\item{PHARMACEUTICAL}{is the chemical used this way, 1 for yes, 0 for no}
\item{INERT}{is the chemical used this way, 1 for yes, 0 for no}
\item{FLAME.RETARDANT}{is the chemical used this way, 1 for yes, 0 for no}
\item{TOXIN}{is the chemical used this way, 1 for yes, 0 for no}
}
}
\usage{
UseTable
}
\description{
A dataset containing 17 indicator columns for chemical uses
such as pesticide, antimicrobial, etc.
}
\keyword{datasets}
