% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Wtns}
\alias{Wtns}
\title{Population distribution of body weight}
\format{
A numberic matrix with 7517 rows and 3 columns
}
\usage{
Wtns
}
\description{
Population distribution of body weight
}
\keyword{datasets}
